Tooltips appear over group parent
Previously, the tooltips would jump around as the user moused over
different elements of the room: the shading, the circle, the pie
slice, and the text.  This was irritating and and slowed down the
interface.

The new behavior locks the tooltip to the group element which is
the parent of all said elements.

Note that this implementation is fragile, as it assumes that all
elements in the room are direct children of the group element. If
any nesting was performed in a room, this functionality would break.